Add files via upload
This script is used to split the data set into train, test and validation set. The validation set is is taken out from a small percentage of test set. This script requires that we import our config settings and paths for collecting all the image paths which is to be saved in a folder named main.